requirements.txt
discord.py
requests
SQLAlchemy
